app/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
scrapy==2.8.0
django-environ==0.10.0
python-dotenv==1.0.0
